Shinobi::generateRules() now includes a ninja rule.
Given the P.I.T.A. that is generating these in cmd.exe compared to bourne,
it might be useful to have a function that just generates a rule like this.
Or one that has a pair of placeholders that can be set accordingly.

Such a function would be needed for a case like javac -d to get it right(tm).
And it'd probably be nice if ngen could choose win vs nix syntax as a
runtime flag instead of always matching platform.