javac backend: kludge for javac -d flag.
$builddir means foo/Bar.java becomes $builddir/Bar.class instead of the
desired $builddir/foo/Bar.class.

This changes it to use $(dirname $out), or the cmd.exe equiv based on the
_WIN32 macro. Better would be a Bundle field that specifies this, a command
line flag to set this, and a default based on the macro.

Because javac and cxxbase/gcc aren't really platform specific the way that
msvc is. Utility functions for the dirname/basename thing would be nice too.