Go variable crazy.
This creates $archive, $library, and $runtime vars kinda like the install()
command in CMake. Because that's a nice way of dealing with the .dll / .lib
issue for Windows.

This also creates GNU autoconf style vars that essentially define the
GNU standard layout that anyone who has ever suffered unix or the GNU
toolchain will be familar with.

Defaults are loaded into a Bundle::data field called "distribution". If
/project/distribution/field exists, Shinobi overrides the /distribution/field
value with it.

cxxbase now uses $bindir and $libdir for its rules, and ./ngen.json resets
this so that it's still generating its lone binary in dist/ instead of
dist/bin.

Further thinking required for how msvc will need to decide on putting the
.dll in $bindir/$runtime and the import lib in $libdir/$library. Along with
the handling the other trappings of Windows DLL files.