Add a means of installing headers.
headers -> array of dirs to recursively install.

header_strip_prefix -> lets you do "include/myproject" in your headers[],
and get $includedir/myproject for the install target.

This generates build statements with the copy rule for each file found.

I've made a ls() function to help with this, but it requires std::filesystem.
Which means C++17, or more hackwittery for older compilers like GCC before
version 7.

For now this just generates an error if you don't have it. I'm not in the
mood to spend my Saturday morning writing an #if win32/#else unix dirent
version when C++17 makes this so terse to write.