Allow exporting variables from package types.
Such that examples/package/ngen.json can define a variable like $msvc_exceptions
that can be used in examples/pacakge/nested/ngen.json to provide the /EHsc
flag.

For simplicity sake the field is just "variables", and an array of strings
that aare meant to be entire lines. E.g. [ "name = value", ... ]