Take a shot at better handling of install and top level targets.
Shinobi now has generate* methods for install and targetName goals.

cxxbase now has helper functions for handling the headers, and includes
them in the install/targetName goals. Code could be better but it's late,
and based on splitting up and extending what was already there.

msvc now makes use of this so that the handling of the .dll/.lib/.exp files
are now saner instead of out of sync with the targetName goal.

Dep handling between projects still needs betterizing for cases like
cxx_application depends on cxx_library's headers, but that's already known.