Fix string versus float issue in NGEN_VERSION.
Adding the extra decimal point broke things 'cuz ostream sees this as
a float rather than a string.

Changed the ngen.json file to use escape a pair of quotes from json, then
from the shell, so that the macro expansion is as a string rather than
the raw value.