Sprinkle pool = console, restat, and generator where it makes sense.
pool = console is added to rules generated for running ninja, make, cmake,
and general execs. These make a better output than waiting for them to run,
and their output being collected for later dumping.

restat and generator are set for much the same.

restat says go stat the outputs again, so if they don't change then you
don't need to get rebuildy crazy.

generator says don't clean by default and doesn't rerun if the args change.
This is logical for make, where you're unlikely to handle any of that well,
and in the case of cmake the generated file reruns cmake if need be.