Add /project/install_files target to install arbitrary files.
Could be a little better but should work.

If there is an install_files field, it is expected to be an array of
objects. Mandatory "input" and "output" strings that are interpreted
relative to $sourcedir/ and $distdir/. Optional "executable" boolean
that defaults to false.

For now it flicks between install and copy targets based on the
executable bit. Which should be enough for simply pushing files in as
part of the install.

I've updated cxx_library with a README.md file to use as a test /
example of the feature.

Better IMHO would be to have the default definition of install and copy
rules built into Shinobi::generateRules() instead of relying on
implementations to do it. Most do but not all.